qmk lib: fix wrong argument type

+++ qmk_keyboards/lib/via_custom_menus.c
@@ -395,11 +395,11 @@ static uint32_t defer_eeprom_update_callback(uint32_t trigger_time, defer_eeprom
 
 // export functions
 
-void defer_eeprom_update_byte(uint8_t channel_id, uint8_t value_id, void *adrs, uint32_t value) {
+void defer_eeprom_update_byte(uint8_t channel_id, uint8_t value_id, void *adrs, uint8_t value) {
   defer_eeprom_update((channel_id << 8) + value_id, BYTE, adrs, value);
 }
 
-void defer_eeprom_update_word(uint8_t channel_id, uint8_t value_id, void *adrs, uint32_t value) {
+void defer_eeprom_update_word(uint8_t channel_id, uint8_t value_id, void *adrs, uint16_t value) {
   defer_eeprom_update((channel_id << 8) + value_id, WORD, adrs, value);
 }
