+++ qmk_keyboards/lib/via_custom_menus.h
@@ -105,6 +105,6 @@ void via_custom_td_save(uint8_t td_index);
 void via_custom_non_mac_fn_get_value(uint8_t value_id, uint8_t *value_data);
 void via_custom_non_mac_fn_set_value(uint8_t value_id, uint8_t *value_data);
 
-void defer_eeprom_update_byte(uint8_t channel_id, uint8_t value_id, void *adrs, uint32_t value);
-void defer_eeprom_update_word(uint8_t channel_id, uint8_t value_id, void *adrs, uint32_t value);
+void defer_eeprom_update_byte(uint8_t channel_id, uint8_t value_id, void *adrs, uint8_t value);
+void defer_eeprom_update_word(uint8_t channel_id, uint8_t value_id, void *adrs, uint16_t value);
 void defer_eeprom_update_dword(uint8_t channel_id, uint8_t value_id, void *adrs, uint32_t value);
